# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts APP",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of Contacts APP"
  },
  "servers": [
    {
      "url": "https://nodejs-hw-mongodb-7-oin4.onrender.com"
    },
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
...
  "paths": {
    "/contact": {
      "get": {
        "summary": "Get all contacts",
        "description": "Retrieve a list of contacts belonging to the authorized user. Supports pagination, filtering, and sorting.",
        "operationId": "getAllContacts",
        "tags": ["Contacts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number for pagination (default is 1)",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "required": false,
            "description": "Number of contacts per page (default is 10)",
            "schema": {
              "type": "integer",
              "example": 10
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "description": "Field name to sort by (e.g. name, email, createdAt)",
            "schema": {
              "type": "string",
              "example": "name"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "description": "Sort order — ascending (asc) or descending (desc)",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "example": "asc"
            }
          },
          {
            "name": "isFavourite",
            "in": "query",
            "required": false,
            "description": "Filter contacts by favourite status",
            "schema": {
              "type": "boolean",
              "example": true
            }
          },
          {
            "name": "contactType",
            "in": "query",
            "required": false,
            "description": "Filter contacts by type (personal, work, home)",
            "schema": {
              "type": "string",
              "enum": ["personal", "work", "home"],
              "example": "work"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved list of contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Successfully found contacts!"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "example": 35
                        },
                        "page": {
                          "type": "integer",
                          "example": 1
                        },
                        "perPage": {
                          "type": "integer",
                          "example": 10
                        },
                        "totalPages": {
                          "type": "integer",
                          "example": 4
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/contacts"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "404": null,
            "$ref": "#/components/responses/401",
            "description": "No contacts found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Contact not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "summary": "Create a new contact",
        "description": "Create a new contact for the authorized user. Supports uploading a profile photo via multipart/form-data.",
        "operationId": "createContact",
        "tags": ["Contacts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Contact's full name",
                    "example": "John Doe"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "Contact's phone number",
                    "example": 380971234567
                  },
                  "email": {
                    "type": "string",
                    "description": "Contact's email address",
                    "example": "johndoe@email.com"
                  },
                  "contactType": {
                    "type": "string",
                    "description": "Type of contact (personal, work, home)",
                    "enum": ["personal", "work", "home"],
                    "example": "personal"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Mark contact as favourite",
                    "example": false
                  },
                  "photo": {
                    "type": "string",
                    "format": "binary",
                    "description": "Optional contact photo file"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a contact",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a contact!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contacts"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "contacts": {
        "type": "object",
        "required": ["name", "phoneNumber", "contactType"],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Unique contact identifier (MongoDB ObjectId)",
            "example": "66fbd96c5ad9c45b17a9e0d1"
          },
          "name": {
            "type": "string",
            "description": "Contact name",
            "example": "John Doe"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Contact phone number",
            "example": "+380991234567"
          },
          "email": {
            "type": ["string", "null"],
            "description": "Contact email (optional)",
            "example": "john.doe@example.com"
          },
          "photo": {
            "type": ["string", "null"],
            "description": "Photo URL of the contact (Cloudinary or local)",
            "example": "https://res.cloudinary.com/demo/image/upload/v1734000000/contact123.jpg"
          },
          "isFavourite": {
            "type": "boolean",
            "description": "Indicates whether the contact is marked as favorite",
            "default": false,
            "example": true
          },
          "contactType": {
            "type": "string",
            "description": "Type of contact",
            "enum": ["work", "home", "personal"],
            "default": "personal",
            "example": "work"
          },
          "userId": {
            "type": "string",
            "description": "The ID of the user who owns this contact",
            "example": "66fbcd9e5efc2c47c9e02a34"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date when the contact was created",
            "example": "2025-10-15T12:45:23.123Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date when the contact was last updated",
            "example": "2025-10-15T13:10:11.654Z"
          }
        }
      }
    },
    "responses": {
      "400": {
        "description": "Invalid request data",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Bad Request — Invalid data format or missing required fields."
                },
                "details": {
                  "type": "string",
                  "example": "The field \"email\" must be a valid email address."
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized — missing or invalid token",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized — Missing or invalid authentication token."
                },
                "details": {
                  "type": "string",
                  "example": "Please provide a valid Bearer token in the Authorization header."
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Internal Server Error — Something went wrong on the server."
                },
                "details": {
                  "type": "string",
                  "example": "Unexpected error while updating contact."
                }
              }
            }
          }
        }
      }
    }
  }
}